functions/notifier/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
firebase-admin==6.4.0
markdown==3.5.2
bleach==6.1.0
